// src/types/about.ts
export type PagePath =
    | '/team'
    | '/team-details'
    | '/projects'
    | '/projects-carousel'
    | '/project-details'
    | '/testimonials'
    | '/pricing'
    | '/testimonials-carousel'
    | '/pricing-carousel'
    | '/gallery'
    | '/faq'
    | '/services'
    | '/residential-cleaning'
    | '/commercial-cleaning'
    | '/deep-cleaning'
    | '/office-cleaning'
    | '/sanitizing-mopping'
    | '/products-left'
    | '/products-right'
    | '/products'
    | '/product-details'
    | '/cart'
    | '/checkout'
    | '/wishlist'
    | '/sign-up'
    | '/login'
    | '/blog'
    | '/blog-carousel'
    | '/blog-list'
    | '/blog-details';